Create Trading strategies for a portfolio
At the end of every trading day, we obtain data of several stocks and generate images using the data. At first, we input the images into the Tech-CNN model and create buy signals or sell signals for the next trading day. Next, we calculate the number of buy signals and allocate cash evenly to the stocks which we will buy on the next trading day. Then, we buy the stocks shortly before the market close on the next trading day. At last, after we hold the stocks for 5 trading days, we sell the stocks shortly before the
market close on the sixth day. It is a long-only strategy.