Create Trading strategy for portfolio(stop-loss order)
In order to reduce the volatility of the strategy, we set a threshold for stop-loss for the strategy. On the fifth trading day after we buy stocks, we check whether the close price on the fourth trading day is lower than that on the first trading day. If loss < −0.01, we sell the stock shortly before the market close on the fifth trading day instead of the sixth trading day.